- CRAI review points from Refactor PR: 02
